Split the validation module into pieces
