+++ changelog/55.breaking.md
@@ -2,3 +2,6 @@
   Use the newly added `input4mips-validation db` commands to control the database instead.
 - Our databases our now multi-file, one file per file entry. This makes it much easier to track changes to the database.
   To load a database from a directory of files, use [`load_database_file_entries`][input4mips_validation.database.database.load_database_file_entries]
+- Changed the function signature of [`input4mips_validation.cvs.loading.load_cvs`][input4mips_validation.cvs.loading.load_cvs].
+  Now it simply expects the `cv_source` rather than requiring a [`RawCVLoader`][input4mips_validation.cvs.loading_raw.RawCVLoader] or not input.
+- Split the `input4mips_validation.validation` module into pieces
