+++ tests/integration/cli/test_db_create.py
@@ -17,7 +17,6 @@ from typer.testing import CliRunner
 
 from input4mips_validation.cli import app
 from input4mips_validation.cvs.loading import load_cvs
-from input4mips_validation.cvs.loading_raw import get_raw_cvs_loader
 from input4mips_validation.database import (
     Input4MIPsDatabaseEntryFile,
     load_database_file_entries,
@@ -46,7 +45,7 @@ def test_basic(tmp_path):
     """
     Write two files in a tree, then make sure we can create the database
     """
-    cvs = load_cvs(get_raw_cvs_loader(DEFAULT_TEST_INPUT4MIPS_CV_SOURCE))
+    cvs = load_cvs(DEFAULT_TEST_INPUT4MIPS_CV_SOURCE)
 
     # Create ourselves a tree
     tree_root = tmp_path / "netcdf-files"
